requirements.txt
# Core
pyyaml>=6.0
python-dotenv>=1.0
numpy>=1.24
pandas>=2.0
requests>=2.31
websocket-client>=1.6

# Hyperliquid
hyperliquid-python-sdk>=0.9

# Models
timesfm>=1.0.0
torch>=2.0
# Kronos installed from GitHub — see README setup instructions

# Backtest / reporting
matplotlib>=3.7
scipy>=1.11

# Utilities
aiohttp>=3.9
